[ticket/15167] Move style install code to manager
PHPBB3-15167
