+++ phpBB/config/default/container/services_text_formatter.yml
@@ -10,6 +10,7 @@ services:
     text_formatter.data_access:
         class: phpbb\textformatter\data_access
         arguments:
+            - '@bitfield.factory'
             - '@dbal.conn'
             - '%tables.bbcodes%'
             - '%tables.smilies%'
