+++ phpBB/includes/message_parser.php
@@ -43,6 +43,7 @@ if (!class_exists('bbcode'))
 */
 class bbcode_firstpass extends bbcode
 {
+
 	var $message = '';
 	var $warn_msg = array();
 	var $parsed_items = array();
@@ -52,6 +53,8 @@ class bbcode_firstpass extends bbcode
 	*/
 	function parse_bbcode()
 	{
+		global $phpbb_container;
+
 		if (!$this->bbcodes)
 		{
 			$this->bbcode_init();
@@ -60,7 +63,7 @@ class bbcode_firstpass extends bbcode
 		global $user;
 
 		$this->bbcode_bitfield = '';
-		$bitfield = new bitfield();
+		$bitfield = $phpbb_container->get('bitfield.factory')->get();
 
 		foreach ($this->bbcodes as $bbcode_name => $bbcode_data)
 		{
